download best model per langpair, not working currnetly

+++ pipeline/opusmt/download-model.sh
@@ -22,23 +22,28 @@ then
     wget -O ${model_list} "https://raw.githubusercontent.com/Helsinki-NLP/Tatoeba-Challenge/master/models/released-model-results.txt"
     download_url=$(grep -P -m 1 "^${source_lang}-${target_lang}" ${model_list} | cut -f 4) 
     echo "###### Using best ${source_lang}-${target_lang} model ${download_url}"
+    mkdir -p $model_dir
 fi
 
-model_zip=${download_url##*/}
-archive_path="${model_dir}/${model_zip}"
+# If model is not "best", check if it has been already downloaded
+if test -f "${model_dir}/${best_model}"; then
+    echo "Model has already been dowloaded to ${model_dir}/${best_model}"
+else
+    model_zip=${download_url##*/}
+    archive_path="${model_dir}/${model_zip}"
 
-wget -O "${archive_path}" "${download_url}"
+    wget -O "${archive_path}" "${download_url}"
 
-cd ${model_dir}
-unzip -j -o "${archive_path}"
-rm ${archive_path}
+    cd ${model_dir}
+    unzip -j -o "${model_zip}"
+    rm ${model_zip}
 
-model_file=$(ls *.npz)
-vocab_file=$(ls *vocab.yml)
-#Create a soft link for the model with the name that the workflow expects 
-ln -s $model_file ${best_model}
-#Also create a standard name link for the vocab
-ln -s $vocab_file "vocab.yml"
-
+    model_file=$(ls *.npz)
+    vocab_file=$(ls *vocab.yml)
+    #Create a soft link for the model with the name that the workflow expects 
+    ln -s $model_file ${best_model}
+    #Also create a standard name link for the vocab
+    ln -s $vocab_file "vocab.yml"
+fi
 
 echo "###### Done: Downloading and extracting opus mt model"
